+++ stairsplus/common.lua
@@ -35,6 +35,9 @@ stairsplus.register_single = function(category, alternate, info, modname, subnam
 	-- This makes node rotation work on placement
 	def.place_param2 = nil
 
+	-- Darken light sources slightly to make up for their smaller visual size
+	def.light_source = math.max(0, (def.light_source or 0) - 1)
+
 	def.on_place = minetest.rotate_node
 	def.groups = stairsplus:prepare_groups(fields.groups)